Only run the sliding sync background updates on the main database
